Warn about quoted queries
